# openapi-v3.json
{"openapi":"3.0.1","info":{"title":"Music Service API","description":"API для управления музыкальным каталогом с ролевой системой доступа","contact":{"name":"Music Service Team","email":"support@musicservice.com"},"license":{"name":"MIT License","url":"https://opensource.org/licenses/MIT"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Development server"},{"url":"https://api.musicservice.com","description":"Production server"}],"security":[{"basicAuth":[]}],"tags":[{"name":"Admin Music Management","description":"Административное управление музыкой. Требует роль ADMIN"},{"name":"User","description":"Управление пользователями"},{"name":"Admin User Management","description":"Административное управление пользователями. Требует роль ADMIN"},{"name":"Music","description":"Управление музыкой. Доступно для ролей USER и ADMIN"},{"name":"Playlist","description":"Управление плейлистами"}],"paths":{"/api/user":{"get":{"tags":["User"],"description":"Получение всех пользователей","operationId":"findAll","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}},"401":{"description":"Не авторизован","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}},"403":{"description":"Доступ запрещен","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}},"security":[{"basicAuth":[]}]},"post":{"tags":["User"],"summary":"Регистрация нового пользователя","description":"Создание пользователя (регистрация). Роль по умолчанию: USER","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"400":{"description":"Некорректные данные","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"200":{"description":"Пользователь успешно создан","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/playlist":{"get":{"tags":["Playlist"],"description":"Получение всех плейлистов","operationId":"findAll_1","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlaylistResponse"}}}}}}},"post":{"tags":["Playlist"],"description":"Создание плейлиста","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaylistRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaylistResponse"}}}}}}},"/api/playlist/music":{"post":{"tags":["Playlist"],"description":"Добавление музыки","operationId":"addMusic","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaylistMusic"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaylistResponse"}}}}}}},"/api/admin/music":{"post":{"tags":["Admin Music Management"],"summary":"Добавить новую песню в каталог","description":"Добавление новой песни. Требует роль ADMIN","operationId":"createMusic","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MusicRequest"}}},"required":true},"responses":{"403":{"description":"Требуется роль ADMIN","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MusicResponse"}}}},"401":{"description":"Не авторизован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MusicResponse"}}}},"200":{"description":"Песня успешно добавлена","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MusicResponse"}}}},"400":{"description":"Некорректные данные","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MusicResponse"}}}}},"security":[{"basicAuth":[]}]}},"/api/user/{id}":{"get":{"tags":["User"],"description":"Получение пользователя по id","operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"401":{"description":"Не авторизован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"403":{"description":"Доступ запрещен","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"basicAuth":[]}]}},"/api/playlist/{id}":{"get":{"tags":["Playlist"],"description":"Получение плейлиста по идентификатору","operationId":"findById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaylistResponse"}}}}}}},"/api/music":{"get":{"tags":["Music"],"summary":"Список всех песен","description":"Получение всех песен. Доступно для ролей USER и ADMIN","operationId":"findAll_2","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MusicResponse"}}}}},"401":{"description":"Не авторизован","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MusicResponse"}}}}},"403":{"description":"Доступ запрещен","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MusicResponse"}}}}}},"security":[{"basicAuth":[]}]}},"/api/music/{id}":{"get":{"tags":["Music"],"summary":"Информация о песне","description":"Получение песни по идентификатору. Доступно для ролей USER и ADMIN","operationId":"findById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MusicResponse"}}}},"401":{"description":"Не авторизован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MusicResponse"}}}},"404":{"description":"Песня не найдена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MusicResponse"}}}},"403":{"description":"Доступ запрещен","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MusicResponse"}}}}},"security":[{"basicAuth":[]}]}},"/api/admin/users":{"get":{"tags":["Admin User Management"],"summary":"Список всех пользователей","description":"Получение всех пользователей. Требует роль ADMIN","operationId":"getAllUsers","responses":{"403":{"description":"Требуется роль ADMIN","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}},"401":{"description":"Не авторизован","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}},"security":[{"basicAuth":[]}]}},"/api/admin/users/{id}":{"get":{"tags":["Admin User Management"],"summary":"Информация о пользователе","description":"Получение пользователя по ID. Требует роль ADMIN","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Требуется роль ADMIN","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"404":{"description":"Пользователь не найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"401":{"description":"Не авторизован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"basicAuth":[]}]},"delete":{"tags":["Admin User Management"],"summary":"Удалить пользователя","description":"Удаление пользователя. Требует роль ADMIN","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Требуется роль ADMIN"},"404":{"description":"Пользователь не найден"},"401":{"description":"Не авторизован"},"200":{"description":"Пользователь успешно удален"}},"security":[{"basicAuth":[]}]}},"/api/admin/music/{id}":{"delete":{"tags":["Admin Music Management"],"summary":"Удалить песню из каталога","description":"Удаление песни. Требует роль ADMIN","operationId":"deleteMusic","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Требуется роль ADMIN"},"200":{"description":"Песня успешно удалена"},"401":{"description":"Не авторизован"},"404":{"description":"Песня не найдена"}},"security":[{"basicAuth":[]}]}}},"components":{"schemas":{"UserRequest":{"type":"object","properties":{"fullName":{"type":"string","description":"Имя"},"address":{"type":"string","description":"Адрес"},"phone":{"type":"string","description":"Номер телефона"},"login":{"type":"string","description":"Логин"},"password":{"type":"string","description":"Пароль"}},"description":"Создание пользователя"},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","description":"Идентификатор","format":"int32"},"fullName":{"type":"string","description":"Имя"},"address":{"type":"string","description":"Адрес"},"phone":{"type":"string","description":"Номер телефона"},"login":{"type":"string","description":"Логин"}},"description":"Пользователь"},"PlaylistRequest":{"type":"object","properties":{"name":{"type":"string","description":"Наименование плейлиста"},"userId":{"type":"integer","description":"Идентификатор пользователя","format":"int64"}},"description":"Создание плейлиста"},"MusicResponse":{"type":"object","properties":{"inventoryNumber":{"type":"string","description":"Инвентарный номер"},"genre":{"type":"string","description":"Жанр"},"title":{"type":"string","description":"Наименование"},"performer":{"type":"string","description":"Исполнитель"},"composerName":{"type":"string","description":"Композитор"},"mediaTypeName":{"type":"string","description":"Тип устройства"},"recordLabelName":{"type":"string","description":"Студия записи"},"recordingDate":{"type":"string","description":"Дата записи","format":"date"}},"description":"Музыка"},"PlaylistResponse":{"type":"object","properties":{"id":{"type":"integer","description":"Идентификатор","format":"int32"},"name":{"type":"string","description":"Наименование"},"user":{"$ref":"#/components/schemas/UserResponse"},"musicList":{"type":"array","description":"Список музыки","items":{"$ref":"#/components/schemas/MusicResponse"}}},"description":"Плейлист"},"PlaylistMusic":{"type":"object","properties":{"playlistId":{"type":"integer","description":"Идентификатор плейлиста","format":"int64"},"inventoryNumbers":{"type":"array","description":"Список идентификаторов музыки","items":{"type":"integer","description":"Список идентификаторов музыки","format":"int64"}}},"description":"Добавление музыки в плейлист"},"MusicRequest":{"type":"object","properties":{"genreId":{"type":"integer","format":"int32"},"title":{"type":"string"},"performerId":{"type":"integer","format":"int32"},"composerId":{"type":"integer","format":"int32"},"mediaTypeId":{"type":"integer","format":"int32"},"recordLabelId":{"type":"integer","format":"int32"},"recordingDate":{"type":"string","format":"date"}}}},"securitySchemes":{"basicAuth":{"type":"http","description":"HTTP Basic Authentication. Используйте логин и пароль для входа.","scheme":"basic"}}}}