# src/main/java/ru/dip4rip/musicservice/models/PlaylistMusic.java
package ru.dip4rip.musicservice.models;

import jakarta.persistence.EmbeddedId;
import jakarta.persistence.Entity;
import jakarta.persistence.FetchType;
import jakarta.persistence.JoinColumn;
import jakarta.persistence.ManyToOne;
import jakarta.persistence.MapsId;
import jakarta.persistence.Table;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@Entity
@Table(name = "playlist_music")
public class PlaylistMusic {
  @EmbeddedId
  private PlaylistMusicId id;

  @MapsId("playlistId")
  @ManyToOne(fetch = FetchType.LAZY, optional = false)
  @JoinColumn(name = "playlist_id", nullable = false)
  private Playlist playlist;

  @MapsId("inventoryNumber")
  @ManyToOne(fetch = FetchType.LAZY, optional = false)
  @JoinColumn(name = "inventory_number", nullable = false)
  private Music music;

}